Fix useless test in longst collatz

+++ project_euler/014/longest_collatz.py
@@ -12,7 +12,7 @@ cache = {
 
 def next_chain(n: int) -> int:
     if n % 2 == 0:
-        return n / 2
+        return int(n / 2)
     else:
         return 3 * n + 1
